Added direct messaging.

+++ items/views.py
@@ -4,8 +4,10 @@ from django.views.generic import ListView, DetailView
 from django.contrib import messages
 from .models import Item, Image
 from .forms import PostItem, ImageFormSet
+from messaging.forms import MessageForm
 from profiles.models import Profile
 from django.db.utils import IntegrityError
+from directmessages.apps import Inbox
 
 # Create your views here.
 
@@ -19,43 +21,59 @@ class ItemDetailView(DetailView):
 
 def item_detail_view(req, pk):
     item = get_object_or_404(Item, id=pk)
-    item_form = PostItem(
-        req.POST or None,
-        req.FILES or None,
-        instance = item,
-    )
-    image_form = ImageFormSet(
-        req.POST or None,
-        req.FILES or None,
-        queryset = item.get_images(),
-        
-    )
-    if req.method == 'POST':
-        redir = False
-        if item_form.is_valid():
-            redir = True
-            item_form.save()
-        if image_form.is_valid():
-            for img in item.get_images():
-                if str(img.id) in req.POST:
-                    assert req.POST.get(str(img.id)) == 'on'
-                    img.delete()
-            redir = True
-            imgs = image_form.save(commit=False)
-            for img in imgs:
-                try:
-                    img.save()
-                except IntegrityError:
-                    img.item = item
-                    img.save()
-        if redir:
-            item_form = PostItem()
-            image_form = ImageFormSet()
-            return HttpResponseRedirect(str(pk))
+    from_user = req.user
+    to_user = item.seller.user
+    is_own_item = from_user == to_user 
+    item_form, image_form, message_form = None, None, None
+    if is_own_item:
+        item_form = PostItem(
+            req.POST or None,
+            req.FILES or None,
+            instance = item,
+        )
+        image_form = ImageFormSet(
+            req.POST or None,
+            req.FILES or None,
+            queryset = item.get_images(),
+            
+        )
+        if req.method == 'POST':
+            redir = False
+            if item_form.is_valid():
+                redir = True
+                item_form.save()
+            if image_form.is_valid():
+                for img in item.get_images():
+                    if str(img.id) in req.POST:
+                        assert req.POST.get(str(img.id)) == 'on'
+                        img.delete()
+                redir = True
+                imgs = image_form.save(commit=False)
+                for img in imgs:
+                    try:
+                        img.save()
+                    except IntegrityError:
+                        img.item = item
+                        img.save()
+            if redir:
+                item_form = PostItem()
+                image_form = ImageFormSet()
+                return HttpResponseRedirect(str(pk))
+    else:
+        message_form = MessageForm()
+        if req.method == 'POST':
+            form_data = {'content': req.POST['content']}
+            message_form = MessageForm(form_data)
+            if message_form.is_valid():
+                message_form.save(from_user, to_user)
     context = {
         'object': item,
         'item_form': item_form,
+        'message_form': message_form,
         'image_form': image_form,
+        'is_own_item': is_own_item,
+        'from_user': from_user,
+        'to_user': to_user,
     }
     return render(req, 'items/detail.html', context)
